Upgrade halide_type_t to 64-bit to support very large vectors and types

+++ src/Type.h
@@ -310,7 +310,7 @@ public:
      * bits: The bit size of one element.
      * lanes: The number of vector elements in the type. */
     Type(halide_type_code_t code, int bits, int lanes, const halide_handle_cplusplus_type *handle_type = nullptr)
-        : type(code, (uint8_t)bits, (uint16_t)lanes), handle_type(handle_type) {
+        : type(code, (uint16_t)bits, (int32_t)lanes), handle_type(handle_type) {
         user_assert(lanes == type.lanes)
             << "Halide only supports vector types with up to 65535 lanes. " << lanes << " lanes requested.";
         user_assert(bits == type.bits)
